some progress 25/04/2022

+++ backend/foodgram/api/views.py
@@ -1,17 +1,18 @@
+import rest_framework.permissions
 from django.db import IntegrityError
-from django.db.models import Count
+from django.db.models import Exists, Count, OuterRef, Q
+from django.apps import apps
 from django.http import FileResponse
 from django.shortcuts import get_object_or_404
 from django_filters.rest_framework import DjangoFilterBackend
-from rest_framework import (pagination, permissions, response, status, views,
-                            viewsets)
+from rest_framework import pagination, response, status, views, viewsets
 
 from recipes.models import (Ingredient, Recipe, RecipeFavorite, ShoppingCart,
                             Tag)
 from users.models import Subscription, User
 
 from .filters import IngredientSearchFilter, RecipeFilter
-from .permissions import IsAuthorOrReadOnly
+from .permissions import RecipePermission
 from .serializers import (RecipesCreateSerializer, RecipesSerializer,
                           SubscribeSerializer)
 from .simple_serializers import (IngredientsSerializer,
@@ -41,58 +42,80 @@ class RecipesViewSet(viewsets.ModelViewSet):
 
     queryset = Recipe.objects.all()
     serializer_class = RecipesSerializer
+    permission_classes = (RecipePermission, )
     pagination_class = pagination.LimitOffsetPagination
     filter_backends = (DjangoFilterBackend, )
     filterset_class = RecipeFilter
 
     def get_queryset(self):
         user = self.request.user
+        if not user.is_authenticated:
+            return self.queryset
 
-        if user.is_anonymous:
-            return Recipe.objects.all()
+        active_filters = []
 
         fav_param = self.request.query_params.get('is_favorited', '0')
-        is_fav = IsFavOrInShopCart(
-            fav_param,
-            'is_favorited'
-        )
+        is_fav = IsFavOrInShopCart(fav_param, 'is_favorited')
+        if is_fav.check:
+            active_filters.append(Q(recipe_favorite__user=user))
+
         cart_param = self.request.query_params.get('is_in_shopping_cart', '0')
-        in_shop_cart = IsFavOrInShopCart(
-            cart_param,
-            'is_in_shopping_cart'
-        )
+        in_shop_cart = IsFavOrInShopCart(cart_param, 'is_in_shopping_cart')
+        if in_shop_cart.check:
+            active_filters.append(Q(in_shopping_cart__user=user))
 
-        if not is_fav.check and not in_shop_cart.check:
-            return Recipe.objects.all()
-
-        if is_fav.check and not in_shop_cart.check:
-            return Recipe.objects.filter(
-                recipe_favorite__user=user
-            ).all()
-
-        if not is_fav.check and in_shop_cart.check:
-            return Recipe.objects.filter(
-                in_shopping_cart__user=user
-            ).all()
-
-        if is_fav and in_shop_cart:
-            return Recipe.objects.filter(
-                in_shopping_cart__user=user
-            ).filter(
-                recipe_favorite__user=user
-            ).all()
-
-    def get_permissions(self):
-        if self.action in ('list', 'retrieve'):
-            self.permission_classes = [permissions.AllowAny, ]
-        elif self.action in ('partial_update', 'destroy'):
-            self.permission_classes = [IsAuthorOrReadOnly, ]
-        return super().get_permissions()
+        return self.queryset.filter(*active_filters)
 
     def get_serializer_class(self):
         if self.action in ('create', 'partial_update'):
             return RecipesCreateSerializer
         return RecipesSerializer
+
+
+class AddToFavOrShopCartCommonView(views.APIView):
+    """
+    Общий APIView для обработки запросов на
+    добавление рецепта в избранное или корзину.
+    """
+
+    def post(self, request, id, primary, secondary):
+        primary = apps.get_model(primary['app'], primary['model'])  #модель связка
+        secondary = apps.get_model(secondary['app'], secondary['model']) #модель рецепт
+        secondary_unit = get_object_or_404(secondary, id=id) #конкретный рецепт
+        # data = {k: v for k, v in zip([f.name for f in primary._meta.get_fields()[1:]], [secondary_unit, request.user])}
+        try:
+            primary.objects.create(
+                recipe=secondary_unit,
+                user=request.user
+            )
+        except IntegrityError as error:
+            return response.Response(
+                data={'errors': str(error.__context__)},
+                status=status.HTTP_400_BAD_REQUEST
+            )
+        serializer = RecipesShortInfoSerializer(secondary_unit)
+        return response.Response(
+            serializer.data,
+            status=status.HTTP_201_CREATED
+        )
+
+    def delete(self, request, id, primary, secondary):
+        primary = apps.get_model(primary['app'], primary['model'])  #модель связка
+        secondary = apps.get_model(secondary['app'], secondary['model']) #модель рецепт
+        secondary_unit = get_object_or_404(secondary, id=id) #конкретный рецепт
+        try:
+            if not primary.objects.filter(recipe=secondary_unit, user=request.user).exists():
+                raise Exception('Этого рецепта нет в избранном/корзине.')
+            primary.objects.filter(
+                recipe=secondary_unit,
+                user=request.user
+            ).delete()
+            return response.Response(status=status.HTTP_204_NO_CONTENT)
+        except Exception as error:
+            return response.Response(
+                data={'errors': str(error)},
+                status=status.HTTP_400_BAD_REQUEST
+            )
 
 
 class MakeSubscription(views.APIView):
@@ -103,7 +126,7 @@ class MakeSubscription(views.APIView):
         try:
             Subscription.objects.create(
                 author=author,
-                subscriber=request.user
+                user=request.user
             )
         except IntegrityError as error:
             return response.Response(
@@ -121,11 +144,19 @@ class MakeSubscription(views.APIView):
 
     def delete(self, request, id):
         author = get_object_or_404(User, id=id)
-        Subscription.objects.filter(
-            author=author,
-            subscriber=request.user
-        ).delete()
-        return response.Response(status=status.HTTP_204_NO_CONTENT)
+        try:
+            if not Subscription.objects.filter(author=author, user=request.user).exists():
+                raise Exception('Такой подписки не существует.')
+            Subscription.objects.filter(
+                author=author,
+                user=request.user
+            ).delete()
+            return response.Response(status=status.HTTP_204_NO_CONTENT)
+        except Exception as error:
+            return response.Response(
+                data={'errors': str(error)},
+                status=status.HTTP_400_BAD_REQUEST
+            )
 
 
 class ShowSubscriptionViewSet(ListViewSet):
@@ -137,7 +168,7 @@ class ShowSubscriptionViewSet(ListViewSet):
     def get_queryset(self):
         user = self.request.user
         return User.objects.filter(
-            subscribers__subscriber=user
+            subscribers__user=user
         ).annotate(
             recipes_count=Count('recipes')
         )
@@ -213,20 +244,19 @@ class DownloadShoppingCart(views.APIView):
     Обработка запроса на скачивание списка покупок.
     После обработки корзина очищается.
     """
-
+    permission_classes = [rest_framework.permissions.AllowAny, ]
     def get(self, request):
         user = request.user
         queryset = user.shopping_cart.select_related('recipe').all()
+        # queryset = ShoppingCart.objects.all()
         message = get_header_message(queryset)
         total_list = get_total_list(queryset)
 
-        with open('total_list', 'w', encoding='utf-8') as f:
+        with open('total_list.txt', 'w', encoding='utf-8') as f:
             f.write(f'{message}\n\n')
             for k, v in total_list.items():
                 for unit, amount in v.items():
                     f.write(f'{k}: {amount} {unit}\n')
 
         user.shopping_cart.all().delete()
-
-        response = FileResponse(open('total_list', 'rb'), as_attachment=True)
-        return response
+        return FileResponse(open('total_list.txt', 'rb'), as_attachment=True)
